requirements_txt.txt
streamlit==1.28.1
tensorflow==2.13.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
transformers==4.33.2
torch==2.0.1
plotly==5.15.0
wordcloud==1.9.2
textstat==0.7.3
arabic-reshaper==3.0.0
python-bidi==0.4.2
Pillow==10.0.0
requests==2.31.0